+++ stubs/.phpstorm.meta.php
@@ -7,12 +7,15 @@
 namespace PHPSTORM_META {
     registerArgumentsSet('CborEncodeFlags',
 /* encode flags start */
-CBOR_SELF_DESCRIBE | CBOR_BYTE | CBOR_TEXT | CBOR_INT_KEY | CBOR_KEY_BYTE | CBOR_KEY_TEXT | CBOR_UNSAFE_TEXT | CBOR_FLOAT16 | CBOR_FLOAT32
+CBOR_SELF_DESCRIBE | CBOR_BYTE | CBOR_TEXT | CBOR_INT_KEY | CBOR_KEY_BYTE
+ | CBOR_KEY_TEXT | CBOR_UNSAFE_TEXT | CBOR_FLOAT16 | CBOR_FLOAT32
 /* encode flags end */
     );
     registerArgumentsSet('CborDecodeFlags',
 /* decode flags start */
-CBOR_SELF_DESCRIBE | CBOR_BYTE | CBOR_TEXT | CBOR_INT_KEY | CBOR_KEY_BYTE | CBOR_KEY_TEXT | CBOR_UNSAFE_TEXT | CBOR_MAP_AS_ARRAY | CBOR_MAP_NO_DUP_KEY | CBOR_FLOAT16 | CBOR_FLOAT32
+CBOR_SELF_DESCRIBE | CBOR_BYTE | CBOR_TEXT | CBOR_INT_KEY | CBOR_KEY_BYTE
+ | CBOR_KEY_TEXT | CBOR_UNSAFE_TEXT | CBOR_MAP_AS_ARRAY | CBOR_MAP_NO_DUP_KEY
+ | CBOR_FLOAT16 | CBOR_FLOAT32 | CBOR_EDN
 /* decode flags end */
     );
 
@@ -30,6 +33,11 @@ CBOR_SELF_DESCRIBE | CBOR_BYTE | CBOR_TEXT | CBOR_INT_KEY | CBOR_KEY_BYTE | CBOR
         'max_size' => 65536,
         'string_ref' => true,
         'shared_ref' => false,
+
+        'indent' => false,
+        'pad' => true,
+        'byte_pad' => 0,
+        'byte_wrap' => false,
     ]);
 
     expectedArguments(\cbor_encode(), 1, argumentsSet('CborEncodeFlags'));